Add 2 access options to Config and refactor DriverFactory
This update adds 2 getters for existing configuration options in `Config`:
- `routingTablePurgeDelayMillis()` - provides access to value configured via `ConfigBuilder.withRoutingTablePurgeDelay(long, TimeUnit)`
- `maxTransactionRetryTimeMillis()` - provides access to value configured via `ConfigBuilder.withMaxTransactionRetryTime(long, TimeUnit)`

It also includes internal `DriverFactory` refactoring.
