+++ driver/src/main/java/org/neo4j/driver/Config.java
@@ -36,7 +36,7 @@ import org.neo4j.driver.internal.SecuritySettings;
 import org.neo4j.driver.internal.async.pool.PoolSettings;
 import org.neo4j.driver.internal.cluster.RoutingSettings;
 import org.neo4j.driver.internal.handlers.pulln.FetchSizeUtil;
-import org.neo4j.driver.internal.retry.RetrySettings;
+import org.neo4j.driver.internal.retry.ExponentialBackoffRetryLogic;
 import org.neo4j.driver.net.ServerAddressResolver;
 import org.neo4j.driver.util.Experimental;
 import org.neo4j.driver.util.Immutable;
@@ -75,7 +75,9 @@ public final class Config implements Serializable {
 
     private static final Config EMPTY = builder().build();
 
-    /** User defined logging */
+    /**
+     * User defined logging
+     */
     private final Logging logging;
 
     private final boolean logLeakedSessions;
@@ -90,9 +92,9 @@ public final class Config implements Serializable {
 
     private final long fetchSize;
     private final long routingTablePurgeDelayMillis;
+    private final long maxTransactionRetryTimeMillis;
 
     private final int connectionTimeoutMillis;
-    private final RetrySettings retrySettings;
     private final ServerAddressResolver resolver;
 
     private final int eventLoopThreads;
@@ -113,7 +115,7 @@ public final class Config implements Serializable {
 
         this.connectionTimeoutMillis = builder.connectionTimeoutMillis;
         this.routingTablePurgeDelayMillis = builder.routingTablePurgeDelayMillis;
-        this.retrySettings = builder.retrySettings;
+        this.maxTransactionRetryTimeMillis = builder.maxTransactionRetryTimeMillis;
         this.resolver = builder.resolver;
         this.fetchSize = builder.fetchSize;
 
@@ -123,6 +125,7 @@ public final class Config implements Serializable {
 
     /**
      * Logging provider
+     *
      * @return the Logging provider to use
      */
     public Logging logging() {
@@ -212,18 +215,21 @@ public final class Config implements Serializable {
     }
 
     /**
-     * @return the security setting to use when creating connections.
+     * Returns stale routing table purge delay.
+     *
+     * @return routing table purge delay
      */
-    SecuritySettings securitySettings() {
-        return securitySettings;
+    public long routingTablePurgeDelayMillis() {
+        return routingTablePurgeDelayMillis;
     }
 
-    RoutingSettings routingSettings() {
-        return new RoutingSettings(routingTablePurgeDelayMillis);
-    }
-
-    RetrySettings retrySettings() {
-        return retrySettings;
+    /**
+     * Returns managed transactions maximum retry time.
+     *
+     * @return maximum retry time
+     */
+    public long maxTransactionRetryTimeMillis() {
+        return maxTransactionRetryTimeMillis;
     }
 
     public long fetchSize() {
@@ -265,9 +271,9 @@ public final class Config implements Serializable {
         private String userAgent = format("neo4j-java/%s", driverVersion());
         private final SecuritySettings.SecuritySettingsBuilder securitySettingsBuilder =
                 new SecuritySettings.SecuritySettingsBuilder();
-        private long routingTablePurgeDelayMillis = RoutingSettings.DEFAULT.routingTablePurgeDelayMs();
+        private long routingTablePurgeDelayMillis = RoutingSettings.STALE_ROUTING_TABLE_PURGE_DELAY_MS;
         private int connectionTimeoutMillis = (int) TimeUnit.SECONDS.toMillis(30);
-        private RetrySettings retrySettings = RetrySettings.DEFAULT;
+        private long maxTransactionRetryTimeMillis = ExponentialBackoffRetryLogic.DEFAULT_MAX_RETRY_TIME_MS;
         private ServerAddressResolver resolver;
         private MetricsAdapter metricsAdapter = MetricsAdapter.DEV_NULL;
         private long fetchSize = FetchSizeUtil.DEFAULT_FETCH_SIZE;
@@ -331,7 +337,7 @@ public final class Config implements Serializable {
          * validity and negative values mean connections will never be tested.
          *
          * @param value the minimum idle time
-         * @param unit the unit in which the duration is given
+         * @param unit  the unit in which the duration is given
          * @return this builder
          */
         public ConfigBuilder withConnectionLivenessCheckTimeout(long value, TimeUnit unit) {
@@ -356,7 +362,7 @@ public final class Config implements Serializable {
          * checked.
          *
          * @param value the maximum connection lifetime
-         * @param unit the unit in which the duration is given
+         * @param unit  the unit in which the duration is given
          * @return this builder
          */
         public ConfigBuilder withMaxConnectionLifetime(long value, TimeUnit unit) {
@@ -402,7 +408,7 @@ public final class Config implements Serializable {
          * of {@code 0} is allowed and results in no timeout and immediate failure when connection is unavailable.
          *
          * @param value the acquisition timeout
-         * @param unit the unit in which the duration is given
+         * @param unit  the unit in which the duration is given
          * @return this builder
          * @see #withMaxConnectionPoolSize(int)
          */
@@ -418,6 +424,7 @@ public final class Config implements Serializable {
 
         /**
          * Set to use encrypted traffic.
+         *
          * @return this builder
          */
         public ConfigBuilder withEncryption() {
@@ -427,6 +434,7 @@ public final class Config implements Serializable {
 
         /**
          * Set to use unencrypted traffic.
+         *
          * @return this builder
          */
         public ConfigBuilder withoutEncryption() {
@@ -461,13 +469,12 @@ public final class Config implements Serializable {
          * The routing table of a database get refreshed if the database is used frequently.
          * If the database is not used for a long time,
          * the driver use the timeout specified here to purge the stale routing table.
-         *
+         * <p>
          * After a routing table is removed, next time when using the database of the purged routing table,
          * the driver will fall back to use seed URI for a new routing table.
-         * @param delay
-         *         the amount of time to wait before purging routing tables
-         * @param unit
-         *         the unit in which the duration is given
+         *
+         * @param delay the amount of time to wait before purging routing tables
+         * @param unit  the unit in which the duration is given
          * @return this builder
          */
         public ConfigBuilder withRoutingTablePurgeDelay(long delay, TimeUnit unit) {
@@ -483,15 +490,16 @@ public final class Config implements Serializable {
         /**
          * Specify how many records to fetch in each batch.
          * This config is only valid when the driver is used with servers that support Bolt V4 (Server version 4.0 and later).
-         *
+         * <p>
          * Bolt V4 enables pulling records in batches to allow client to take control of data population and apply back pressure to server.
          * This config specifies the default fetch size for all query runs using {@link Session} and {@link org.neo4j.driver.async.AsyncSession}.
          * By default, the value is set to {@code 1000}.
          * Use {@code -1} to disables back pressure and config client to pull all records at once after each run.
-         *
+         * <p>
          * This config only applies to run result obtained via {@link Session} and {@link org.neo4j.driver.async.AsyncSession}.
          * As with {@link org.neo4j.driver.reactive.RxSession}, the batch size is provided via
          * {@link org.reactivestreams.Subscription#request(long)} instead.
+         *
          * @param size the default record fetch size when pulling records in batches using Bolt V4.
          * @return this builder
          */
@@ -512,10 +520,10 @@ public final class Config implements Serializable {
          * The default value of this parameter is {@code 30 SECONDS}.
          *
          * @param value the timeout duration
-         * @param unit the unit in which duration is given
+         * @param unit  the unit in which duration is given
          * @return this builder
          * @throws IllegalArgumentException when given value is negative or does not fit in {@code int} when
-         * converted to milliseconds.
+         *                                  converted to milliseconds.
          */
         public ConfigBuilder withConnectionTimeout(long value, TimeUnit unit) {
             long connectionTimeoutMillis = unit.toMillis(value);
@@ -534,16 +542,14 @@ public final class Config implements Serializable {
         }
 
         /**
-         * Specify the maximum time transactions are allowed to retry via
-         * {@link Session#readTransaction(TransactionWork)} and {@link Session#writeTransaction(TransactionWork)}
-         * methods. These methods will retry the given unit of work on {@link org.neo4j.driver.exceptions.ServiceUnavailableException},
-         * {@link org.neo4j.driver.exceptions.SessionExpiredException} and {@link org.neo4j.driver.exceptions.TransientException} with
-         * exponential backoff using initial delay of 1 second.
+         * Specify the maximum time managed transactions are allowed to retry.
+         * <p>
+         * Managed transactions are available via methods like {@link Session#executeRead(TransactionCallback)}, {@link Session#executeWrite(TransactionCallback, TransactionConfig)} and some other variations available under similar naming.
          * <p>
          * Default value is 30 seconds.
          *
          * @param value the timeout duration
-         * @param unit the unit in which duration is given
+         * @param unit  the unit in which duration is given
          * @return this builder
          * @throws IllegalArgumentException when given value is negative
          */
@@ -553,7 +559,7 @@ public final class Config implements Serializable {
                 throw new IllegalArgumentException(
                         String.format("The max retry time may not be smaller than 0, but was %d %s.", value, unit));
             }
-            this.retrySettings = new RetrySettings(maxRetryTimeMs);
+            this.maxTransactionRetryTimeMillis = maxRetryTimeMs;
             return this;
         }
 
@@ -586,6 +592,7 @@ public final class Config implements Serializable {
 
         /**
          * Disable driver metrics. When disabled, driver metrics cannot be accessed via {@link Driver#metrics()}.
+         *
          * @return this builder.
          */
         public ConfigBuilder withoutDriverMetrics() {
@@ -619,6 +626,7 @@ public final class Config implements Serializable {
         /**
          * Configure the event loop thread count. This specifies how many threads the driver can use to handle network I/O events
          * and user's events in driver's I/O threads. By default, 2 * NumberOfProcessors amount of threads will be used instead.
+         *
          * @param size the thread count.
          * @return this builder.
          * @throws IllegalArgumentException if the value of the size is set to a number that is less than 1.
@@ -634,6 +642,7 @@ public final class Config implements Serializable {
 
         /**
          * Configure the user_agent field sent to the server to identify the connected client.
+         *
          * @param userAgent the string to configure user_agent.
          * @return this builder.
          */
@@ -802,6 +811,7 @@ public final class Config implements Serializable {
 
         /**
          * The revocation strategy used for verifying certificates.
+         *
          * @return this {@link TrustStrategy}'s revocation strategy
          */
         public RevocationCheckingStrategy revocationCheckingStrategy() {
@@ -811,6 +821,7 @@ public final class Config implements Serializable {
         /**
          * Configures the {@link TrustStrategy} to not carry out OCSP revocation checks on certificates. This is the
          * option that is configured by default.
+         *
          * @return the current trust strategy
          */
         public TrustStrategy withoutCertificateRevocationChecks() {
@@ -823,6 +834,7 @@ public final class Config implements Serializable {
          * stapled to the certificate. If no stapled response is found, then certificate verification continues
          * (and does not fail verification). This setting also requires the server to be configured to enable
          * OCSP stapling.
+         *
          * @return the current trust strategy
          */
         public TrustStrategy withVerifyIfPresentRevocationChecks() {
@@ -834,9 +846,10 @@ public final class Config implements Serializable {
          * Configures the {@link TrustStrategy} to carry out strict OCSP revocation checks for revocation status that
          * are stapled to the certificate. If no stapled response is found, then the driver will fail certificate verification
          * and not connect to the server. This setting also requires the server to be configured to enable OCSP stapling.
-         *
+         * <p>
          * Note: enabling this setting will prevent the driver connecting to the server when the server is unable to reach
          * the certificate's configured OCSP responder URL.
+         *
          * @return the current trust strategy
          */
         public TrustStrategy withStrictRevocationChecks() {
