+++ driver/src/test/java/org/neo4j/driver/integration/ErrorIT.java
@@ -57,9 +57,7 @@ import org.neo4j.driver.Session;
 import org.neo4j.driver.Transaction;
 import org.neo4j.driver.exceptions.ClientException;
 import org.neo4j.driver.exceptions.ServiceUnavailableException;
-import org.neo4j.driver.internal.cluster.RoutingSettings;
 import org.neo4j.driver.internal.messaging.response.FailureMessage;
-import org.neo4j.driver.internal.retry.RetrySettings;
 import org.neo4j.driver.internal.security.SecurityPlanImpl;
 import org.neo4j.driver.internal.util.FailingMessageFormat;
 import org.neo4j.driver.internal.util.FakeClock;
@@ -261,13 +259,11 @@ class ErrorIT {
 
         URI uri = session.uri();
         AuthToken authToken = session.authToken();
-        RoutingSettings routingSettings = RoutingSettings.DEFAULT;
-        RetrySettings retrySettings = RetrySettings.DEFAULT;
         Config config = Config.builder().withLogging(DEV_NULL_LOGGING).build();
         Throwable queryError = null;
 
-        try (Driver driver = driverFactory.newInstance(
-                uri, authToken, routingSettings, retrySettings, config, SecurityPlanImpl.insecure())) {
+        try (Driver driver =
+                driverFactory.newInstance(uri, authToken, config, SecurityPlanImpl.insecure(), null, null)) {
             driver.verifyConnectivity();
             try (Session session = driver.session()) {
                 messageFormatSetup.accept(driverFactory.getFailingMessageFormat());
