+++ driver/src/test/java/org/neo4j/driver/integration/SessionIT.java
@@ -83,8 +83,6 @@ import org.neo4j.driver.exceptions.ResultConsumedException;
 import org.neo4j.driver.exceptions.ServiceUnavailableException;
 import org.neo4j.driver.exceptions.TransientException;
 import org.neo4j.driver.internal.DriverFactory;
-import org.neo4j.driver.internal.cluster.RoutingSettings;
-import org.neo4j.driver.internal.retry.RetrySettings;
 import org.neo4j.driver.internal.security.SecurityPlanImpl;
 import org.neo4j.driver.internal.util.DisabledOnNeo4jWith;
 import org.neo4j.driver.internal.util.DriverFactoryWithFixedRetryLogic;
@@ -1275,12 +1273,7 @@ class SessionIT {
     private Driver newDriverWithFixedRetries(int maxRetriesCount) {
         DriverFactory driverFactory = new DriverFactoryWithFixedRetryLogic(maxRetriesCount);
         return driverFactory.newInstance(
-                neo4j.uri(),
-                neo4j.authToken(),
-                RoutingSettings.DEFAULT,
-                RetrySettings.DEFAULT,
-                noLoggingConfig(),
-                SecurityPlanImpl.insecure());
+                neo4j.uri(), neo4j.authToken(), noLoggingConfig(), SecurityPlanImpl.insecure(), null, null);
     }
 
     private Driver newDriverWithLimitedRetries(int maxTxRetryTime, TimeUnit unit) {
