+++ driver/src/test/java/org/neo4j/driver/internal/util/FixedRetryLogic.java
@@ -22,7 +22,6 @@ import static org.neo4j.driver.internal.logging.DevNullLogging.DEV_NULL_LOGGING;
 
 import io.netty.util.concurrent.EventExecutorGroup;
 import org.neo4j.driver.internal.retry.ExponentialBackoffRetryLogic;
-import org.neo4j.driver.internal.retry.RetrySettings;
 
 public class FixedRetryLogic extends ExponentialBackoffRetryLogic {
     private final int retryCount;
@@ -33,7 +32,7 @@ public class FixedRetryLogic extends ExponentialBackoffRetryLogic {
     }
 
     public FixedRetryLogic(int retryCount, EventExecutorGroup eventExecutorGroup) {
-        super(new RetrySettings(Long.MAX_VALUE), eventExecutorGroup, new SleeplessClock(), DEV_NULL_LOGGING);
+        super(Long.MAX_VALUE, eventExecutorGroup, new SleeplessClock(), DEV_NULL_LOGGING);
         this.retryCount = retryCount;
     }
